fix: refactor base processor class into pure functions
