+++ packages/offix-offline/src/index.ts
@@ -23,6 +23,6 @@ export * from "./conflicts/state/ObjectState";
 export * from "./conflicts/strategies/ConflictResolutionStrategy";
 export * from "./conflicts/strategies/ConflictMetaData";
 export * from "./conflicts/state/VersionedState";
-export * from "./conflicts/BaseProcessor";
+export * from "./conflicts/baseHelpers";
 export * from "./conflicts/strategies/ConflictResolutionData";
 export * from "./offline/OfflineError";
